Create note strategy
