+++ plugs/eventfilters/__init__.py
@@ -1,2 +1,3 @@
 
 from .filters import *
+from .index import *
