+++ plugs/mappingstrategies/__init__.py
@@ -9,3 +9,4 @@ from .mappingstrategy import IMappingStrategy
 
 from .pedalstrategy import PedalStrategy
 from .wheelstrategy import WheelStrategy
+from .notestrategy import NoteStrategy
